+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### V0.107
 
 - Removed physics functionality from editor
+- Removed profiling functionality from editor
 
 ### V0.106